MiniMaven: do not accept too-old versions from multi-projects
When we looked harder for a dependency by scanning multi-project
directories (such as Fiji's modules/ directory structure), it is still not
okay for a POM to have an out-dated version.

This happened for example when imglib1 was parsed first, incurring a
dependency on mpicbg version 20111128, and then blockmatching was parsed
which requires mpicbg version 20120621.

This should make Jacquie Jenkins very, very happy.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>